requirements.txt
playwright
crawl4ai
openai
asyncio
psycopg2-binary
pgvector
streamlit
python-dotenv
tiktoken